chore: split tests to separate dir at root
- move tests and fixtures to their relative `./test/` dirs.
- create a `./test/test_helper.h` file for shared methods.
- remove bool return types in favor of int.
- clean up all headers.
- update build/run instructions.
- add `./test/README.md` with list of build/run commands.
