+++ test/base58/fixtures/base58_fixtures.h
