+++ test/ripemd160/fixtures/ripemd160_binary.fixtures
